fix: commit Cargo.lock, enforce --locked, group coupled deps (#3409)
* fix: revert digest 0.11 to 0.10 to restore std feature

Dependabot bumped digest 0.10 to 0.11 but did not bump the dependent
RustCrypto crates (md4, sha1, sha2) which still pin to 0.10. digest
0.11.2 removed the `std` feature, so the resolver fails:

  package `checksums` depends on `digest` with feature `std` but
  `digest` does not have that feature

Reverting digest to 0.10 keeps the RustCrypto trait family aligned
on a single major version. A coordinated bump to 0.11 across digest,
md4, sha1, and sha2 (with `std` removed from features) is a separate
follow-up.

* fix: revert cargo_metadata 0.23 to 0.18 to honor MSRV 1.88

cargo_metadata 0.23.x transitively requires cargo-platform 0.3.x,
which in turn requires rustc 1.91. The workspace MSRV is 1.88
(rust-toolchain.toml), so CI fails dependency resolution with:

  error: rustc 1.88.0 is not supported by the following package:
  cargo-platform@0.3.3 requires rustc 1.91

Revert to cargo_metadata 0.18 which keeps cargo-platform on the
0.1.x line that supports the project MSRV. Bumping cargo_metadata
again should be paired with an MSRV bump in a coordinated change.

* ci: commit Cargo.lock, enforce --locked, group coupled deps

oc-rsync is a binary, so per Rust best practice the Cargo.lock should be
committed. With it gitignored, CI re-resolved dependencies on every run
and individual dependabot bumps to coupled crates silently broke master:

- digest 0.10 -> 0.11 dropped the std feature that md4/sha1/sha2 0.10
  required (commit 563417c0).
- cargo_metadata 0.18 -> 0.23 transitively raised cargo-platform to
  0.3.x which requires rustc 1.91, while MSRV is 1.88 (commit 4d6ce3c7).

Both bumps merged green because dependabot does its own --locked build
and master CI did not. The systemic fix:

1. Track Cargo.lock so dependency resolution is deterministic and
   reviewable. Future bumps must update the lockfile in the same PR,
   making coupling regressions visible at review time.
2. Pass --locked to every cargo build/check/clippy/nextest invocation
   in CI workflows so manifest-vs-lockfile drift fails fast.
3. Group coupled crate families in dependabot:
   - RustCrypto trait family (digest, md-5, md4, sha1, sha2, blake2,
     blake3) bumps together across major/minor/patch.
   - cargo_metadata + cargo-platform bump together so MSRV-bumping
     transitive changes land in one reviewable PR.

The .gitignore comment that hinted Cargo.lock should be tracked for
executables is now realised.

* ci: regenerate Cargo.lock to satisfy --locked

The lockfile committed in 74866233 was out of sync with manifest
constraints. Cargo emitted "the lock file needs to be updated but
--locked was passed to prevent this" on every CI job.

Regenerated via `cargo generate-lockfile`. The pinned versions for
the just-reverted crates remain on the older line: digest 0.10.7,
md-5 0.10.6, md4 0.10.2, sha1 0.10.6, sha2 0.10.9, cargo_metadata
0.18.1, cargo-platform 0.1.9. Going forward, --locked enforces
parity in CI; dependabot grouping keeps coupled crate families
moving together.

* fix: drop const+unsafe from Uid/Gid::from_raw wrappers

rustix 1.1.x changed `Uid::from_raw` and `Gid::from_raw` to safe
non-const fns, so the const+unsafe wrappers in metadata::ownership
no longer compile under -D warnings (unused_unsafe + E0015).

Removing the const and unsafe annotations matches the upstream API
and removes the now-redundant `#![allow(unsafe_code)]` from this
file. The metadata crate still permits unsafe elsewhere per project
policy; this just stops using it where rustix made it unnecessary.

* refactor: tidy ownership wrappers - drop redundant cfg, add inline

Module is already gated by `#[cfg(unix)] mod ownership;` in lib.rs, so
the per-function `#[cfg(unix)]` attributes were dead code. Add `#[inline]`
for cross-crate inlining and clarify the module doc to describe the
actual purpose (cross-platform abstraction parallel to `ownership_stub`)
rather than the obsolete one (encapsulating unsafe, no longer needed
after rustix 1.1's safe `from_raw`).

* fix: rustix 1.1 SeekFrom::Data/Hole take u64 not i64

rustix 1.1.x changed SeekFrom::Data and SeekFrom::Hole variants to
accept u64 instead of i64. The pos and data_start variables were
already u64; the as i64 casts produced an E0308 type mismatch on the
Linux-only sparse hole detection path.

The function is target_os = "linux" gated so the change only affects
Linux builds; non-Linux platforms use detect_holes_fallback and are
unaffected.

* fix: socket2 0.6 renamed set_nodelay/set_tos methods

socket2 0.6 (merged via dependabot PR #3407) renamed `set_nodelay` to
`set_tcp_nodelay` and split the unified `set_tos` into address-family
specific `set_tos_v4(u8)` and `set_tclass_v6(u32)`. This breaks master
CI and blocks all open PRs.

Update apply_socket_options_impl to use the new method names and add
an apply_ip_tos helper that dispatches to v4/v6 based on the socket's
local address family. Mirrors upstream rsync's socket.c semantics where
IP_TOS only succeeds on AF_INET sockets; we apply the equivalent
IPV6_TCLASS for AF_INET6.

* fix: socket2 0.6 / rustix wrapper test fallout

- daemon tests use SockRef::tcp_nodelay() (renamed from .nodelay())
- engine tests drop unsafe wrapper around Uid/Gid::from_raw (no longer
  unsafe after 12257879 dropped const+unsafe from rustix wrappers)

* fix: enable socket2 "all" feature for set_tclass_v6

socket2 0.6 gates platform-specific options (including set_tclass_v6 for
IPV6_TCLASS) behind the "all" cargo feature. Without it, IP_TOS handling
on IPv6 sockets fails to compile on Linux musl, Windows GNU, and the
feature-flag combination matrix.

Mirrors upstream rsync's socket.c which calls setsockopt for IPV6_TCLASS
unconditionally on Unix.

* ci: allowlist GHSA-f5v4-2wr6-hqmg pending russh bump

russh 0.45 has a high-severity pre-auth DoS via unbounded allocation in
the keyboard-interactive auth handler (GHSA-f5v4-2wr6-hqmg). The patch
ships in russh 0.60.1, a 15-minor-version jump that requires an API
rewrite of the embedded SSH module (3000+ lines across 9 files).

The embedded-ssh feature is opt-in and excluded from the default build
set; the vulnerable code path is only reachable when a user explicitly
enables it. Allowlist the advisory in dependency review while the
upgrade work is tracked separately, so the informational check stops
masking real findings on unrelated PRs.

* fix: gate IPV6_TCLASS to Unix - socket2 0.6 only exposes it there

socket2 0.6 only defines `set_tclass_v6` in `sys/unix.rs`, so the
Windows GNU cross-check fails to compile the IPv6 branch of
apply_ip_tos. Upstream rsync's `socket.c` only calls
setsockopt(..., IPPROTO_IP, IP_TOS, ...), which is itself v4-only,
so on Windows we mirror that by skipping the v6 path entirely.

+++ crates/engine/src/local_copy/executor/file/sparse/reader.rs
@@ -92,7 +92,7 @@ impl SparseReader {
         let mut pos = 0u64;
 
         while pos < file_size {
-            match rustix::fs::seek(fd, RustixSeekFrom::Data(pos as i64)) {
+            match rustix::fs::seek(fd, RustixSeekFrom::Data(pos)) {
                 Ok(data_start) => {
                     if data_start > pos {
                         regions.push(SparseRegion::Hole {
@@ -101,7 +101,7 @@ impl SparseReader {
                         });
                     }
 
-                    match rustix::fs::seek(fd, RustixSeekFrom::Hole(data_start as i64)) {
+                    match rustix::fs::seek(fd, RustixSeekFrom::Hole(data_start)) {
                         Ok(hole_start) => {
                             if hole_start > data_start {
                                 regions.push(SparseRegion::Data {
